16.15: Code cleanup -- moving testing function into its own file
