+++ Java/Ch 16. Moderate/Q16_15_Master_Mind/Result.java
@@ -1,0 +1,16 @@
+package Q16_15_Master_Mind;
+
+public class Result {
+	public int hits;
+	public int pseudoHits;
+	
+	public Result(int h, int p) {
+		hits = h;
+		pseudoHits = p;
+	}
+	
+	@Override
+	public String toString() {
+		return "Result [hits=" + hits + ", pseudoHits=" + pseudoHits + "]";
+	}
+};
